scraper_factory/requirements.txt
algoliasearch==3.0.0
boto3==1.34.122
requests==2.31.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9